refactor: reorganize components and remove unused files
- Removed the DappCredentialCard and FlipCredential components as they were no longer needed.
- Refactored the page.tsx file to streamline imports and improve structure by organizing components into feature-based directories.
- Added new components for layout and effects, including ScrollProgress, Footer, Aurora, and StarsBackground.
- Introduced new FAQ and HowItWorks sections to enhance user experience and provide better information.
- Updated the HeroSubtitle and AnimatedActa components for improved visual presentation.
