+++ src/components/effects/Aurora.tsx
