+++ src/features/value-proposition/ValueDetails.tsx
